RenamePolicies Feature
Added the ability for a user to rename policies as they are uploaded into Microsoft Teams.

Maybe you have some weird old school naming convention from the OCS days, perhaps you are moving to a different naming standard as part of the transition to Teams. Simply run the script with the -RenamePolicies flag set and as it uploads each policy it will prompt you for a new name.

Enter the new name or simply press return to accept the existing one.